Warning fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15977 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi_pcap/tapi_pcap.c
@@ -199,7 +199,8 @@ tapi_pcap_pkt_handler(char *fn, void *user_param)
     }
     pkt_len = rc;
     
-    VERB("%s: Packet payload length %ld bytes\n", __FUNCTION__, pkt_len);
+    VERB("%s: Packet payload length %u bytes\n",
+         __FUNCTION__, (unsigned)pkt_len);
 
     pkt = malloc(pkt_len);
     if (pkt == NULL)
